.history/requirements_20250615110234.txt
streamlit
requests
pandas
google-search-results
google-generativeai
python-dotenv
